Fixed atexit bug

+++ agda_comp/src/agda_comp/cli.py
@@ -2,10 +2,13 @@ import argparse
 from pathlib import Path
 import subprocess
 import networkx as nx
+import time
 
 from agda_comp import test_lvl
 
 import atexit
+
+_finished = False
 
 def main():
     parser = argparse.ArgumentParser(description="Fast Agda type checker")
@@ -41,7 +44,7 @@ def agda_compile(module, clean, jobs):
         print("Couldn't find project directory from project file")
 
     # Create dot file if it doesn't exist
-    dot_file = Path("/tmp/agda_tree_graph.dot")
+    dot_file = Path(f"/tmp/agda_tree_graph_{module.stem}.dot")
     
     if not dot_file.exists() or clean:
         print("Creating dot file")
@@ -70,8 +73,6 @@ def agda_compile(module, clean, jobs):
 
     test_lvl.create_test(dot, dir=source_dir, m=jobs)
 
-
-
     commands = "; ".join([
         f"cd {project_dir}/source",
         "cd ..",
@@ -83,14 +84,27 @@ def agda_compile(module, clean, jobs):
     ])
 
     atexit.register(on_exit, project_dir)
+
+    start = time.perf_counter()
     subprocess.run(commands, shell=True)
+    runtime = time.perf_counter() - start
+
+    set_finished()
+
+    print("-"*20)
+    print(f"Compilation took {runtime:.2} seconds")
+
+def set_finished():
+    global _finished
+    _finished = True
 
 # Clean up index files when program exists due to any reason
 def on_exit(project_dir):
-    commands = "; ".join([
-        f"cd {project_dir}",
-        'find ./source -name "index-*lagda" -delete',
-        "rm Makefile",
-    ])
+    if not _finished:
+        commands = "; ".join([
+            f"cd {project_dir}",
+            'find ./source -name "index-*lagda" -delete',
+            "rm Makefile",
+        ])
 
-    subprocess.run(commands, shell=True)
+        subprocess.run(commands, shell=True)
